CMake fixes for gtk3/Linux and gtk2/Mingw builds.

+++ frontend/apps/gtkmm/src/PreferencesDialog.cc
@@ -875,7 +875,7 @@ PreferencesDialog::on_sound_theme_changed()
   TRACE_ENTER("PreferencesDialog::on_sound_theme_changed");
   int idx = sound_theme_button->get_active_row_number();
 
-  if (idx >= 0 && idx < sound_themes.size())
+  if (idx >= 0 && idx < (int)sound_themes.size())
     {
       SoundTheme::Theme &theme = sound_themes[idx];
